reorganizing scripts
